Allow environment variables to be used in lieu of command line flags
When running in Kubernetes or in a Docker cluster, the command line
flags can't be stored outside of the manifest file. Since many companies
store the manifest files in git, we don't want to keep secrets there.

The github.com/namsral/flag go package is a drop in replacement for
the flag package so that environment variables are automatically
checked in the absence of a command line flag.